automation/requirements/req.txt
selenium==3.141.0
robotframework==4.0.3